mcp_integration_example.py
"""
MCP Server Implementation Example for Caption Colorizer
This shows how you could integrate with ChatGPT using MCP if needed
"""

import json
import asyncio
from typing import AsyncGenerator, Optional, Dict, Any
from fastapi import FastAPI, Header, HTTPException
from fastapi.responses import StreamingResponse
import logging

# This is a simplified example - full implementation would need:
# - File download/upload logic
# - S3 or cloud storage integration
# - Actual caption processing integration
# - Authentication and security

app = FastAPI(title="Caption Colorizer MCP Server")

# MCP Tool Definitions that ChatGPT would see
TOOL_DEFINITIONS = [
    {
        "name": "process_captions",
        "description": "Process a video with SRT subtitles to generate colored caption PNGs and Premiere Pro XML",
        "inputSchema": {
            "type": "object",
            "properties": {
                "video_url": {
                    "type": "string",
                    "description": "Direct URL to the video file"
                },
                "srt_content": {
                    "type": "string",
                    "description": "The complete SRT subtitle file content"
                },
                "seed": {
                    "type": "integer",
                    "description": "Optional random seed for reproducible colors"
                }
            },
            "required": ["video_url", "srt_content"]
        }
    }
]

@app.get("/mcp/sse")
async def mcp_sse_endpoint(authorization: Optional[str] = Header(None)):
    """
    Server-Sent Events endpoint for MCP
    ChatGPT connects to this endpoint
    """
    
    async def event_stream() -> AsyncGenerator[str, None]:
        # Initial connection message
        yield f"data: {json.dumps({'type': 'connection', 'status': 'connected'})}\n\n"
        
        # Send available tools
        yield f"data: {json.dumps({'type': 'tools', 'tools': TOOL_DEFINITIONS})}\n\n"
        
        # Keep connection alive
        while True:
            await asyncio.sleep(30)
            yield f"data: {json.dumps({'type': 'ping'})}\n\n"
    
    return StreamingResponse(
        event_stream(),
        media_type="text/event-stream"
    )

# To use with ChatGPT:
# 1. Deploy this server to a public URL
# 2. Configure in ChatGPT's MCP settings:
#    {
#      "type": "url",
#      "url": "https://your-server.com/mcp/sse",
#      "name": "caption-colorizer",
#      "authorization_token": "your-token"
#    }
