requirements.txt
click
PyYAML
tqdm
lxml
pysrt
ffmpeg-python
pytest
Pillow
